pyproject.toml
[project]
name = "ifr-trainer"
version = "0.1.0"
description = "Web-based instrument flying proficiency game for GA pilots"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1",
    "psycopg[binary]>=3.2",
    "pyyaml>=6.0",
    "anthropic>=0.40",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-django",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py"]
python_paths = ["."]
